Added Logs to Marmot :loud_sound::see_no_evil:
This update adds logging to Marmot. Logs are created for both the formatter and and plot_main including all modules. Two types of log files are created, full log of all info and warning statements and a seperate file just for tracking warning messages. The formatting and other config settings for logging can be adjusted in the marmot_logging_config.yml file